+++ edgy/core/marshalls/config.py
@@ -1,14 +1,15 @@
 from __future__ import annotations
 
-from typing import TypedDict
+from typing import TYPE_CHECKING, TypedDict
 
-from edgy.core.db import models  # fixes model not defined errors in pydantic
+if TYPE_CHECKING:
+    from edgy.core.db.models.model import Model
 
 
 class ConfigMarshall(TypedDict, total=False):
     """A TypedDict for configuring Marshall behaviour."""
 
-    model: type[models.Model] | str
+    model: type[Model] | str
     """The model from there the marshall will read from."""
 
     fields: list[str] | None = None