+++ tests/foreign_keys/test_foreignkey_unique_and_no_constraint.py
@@ -285,7 +285,7 @@ async def test_nullable_foreign_key():
     assert member.team is None
 
 
-def test_assertation_error_on_set_null():
+async def test_assertation_error_on_set_null():
     with pytest.raises(FieldDefinitionError) as raised:
 
         class MyModel(edgy.StrictModel):
@@ -297,7 +297,7 @@ def test_assertation_error_on_set_null():
     assert raised.value.args[0] == "When SET_NULL is enabled, null must be True."
 
 
-def test_assertation_error_on_missing_on_delete():
+async def test_assertation_error_on_missing_on_delete():
     with pytest.raises(FieldDefinitionError) as raised:
 
         class MyModel(edgy.StrictModel):