+++ tests/foreign_keys/test_many_to_many_field_special_auto.py
@@ -31,10 +31,6 @@ class Studio(edgy.StrictModel):
         registry = models
 
 
-def test_check_tablename():
-    assert Studio.meta.fields["users"].through.meta.tablename == "fromages_usrs"
-
-
 @pytest.fixture(autouse=True, scope="function")
 async def create_test_database():
     async with database:
@@ -42,6 +38,10 @@ async def create_test_database():
         yield
         if not database.drop:
             await models.drop_all()
+
+
+async def test_check_tablename():
+    assert Studio.meta.fields["users"].through.meta.tablename == "fromages_usrs"
 
 
 async def test_many_to_many_many_fields():